avoid np usage, bug fixes
problem here is that for now there still isn't torch.unique for gpu, so if
we want to compute the loss with data on gpu we get a slowdown, so for now the loss avoid almost completely to put data on cuda. Hopefully we will get a speedup when torch.unique will be available.